STM32: Rename Enable_Clock procedure to Power_Up
Even if the name Enable_Clock is accurate with regards to the
documentation and the internal behavior of the MCU, it is not clear at
all from a user point of view.

This patches use the name Power_Up which describes better the actual use
a these procedures.

Note: This is a partial commit to propose the idea. If it is well
received I will rename the remaining Enable_Clock.