HAL.Audio: Replace Audio_Device with Audio_Stream interface
The current version of HAL.Audio defines an interface Audio_Device that
is actually more a media controller (play, pause, resume volume). it
could be interesting to have a Media_Controller interface at some point
because it's a service that can be provided by different back-ends (USB,
Bluetooth, IR) but right now we don't have any.

This patch remove the "Audio_Device" and replace it with a more
Audio_Stream with three subprograms:

 - Transmit audio buffer
 - Receive audio buffer
 - Set sample frequency