[RFC] Tentative solution for the delay problem
The idea here is that some piece of code is holds an access to an object
(Delay_Provider) that provides delay for the entire library.

The main drawback of course is that someone has to set the
Delay_Provider, in this patch I do it in the elaboration of the board
packages.

I'm not very convince myself about this solution but I though I'd try
and share it.