Renames Services to Middleware
The name services seems to be confusing so I propose to rename it to
middleware. Although the middleware word is quite overloaded in software
development, I think the signification is clear enough for this project.
Also it's the word use by ST for in STM32Cube to hold USB stack, network
stack, filesystems, audio tools, which coresponds with the definition of
what was is currently called services.