STM32F4.I2C: Fix master transmit
Looks like this code was badly copied from the reference drivers. We are
not supposed to wait for BTF to be set, but just check it once. Note
that when reading the documentation, the behavior of this BTF flag is
still not clear to me. I think I spotted this error by reducing the
speed on an I2C port for an experiment. The result was that one out of
two bytes was dropped.