In HAL.Bitmap, add control over the data cache manipulation.
In order to copy big chunks of data from buffer to buffer, it is usually
desirable to use DMA and thus we need to ensure that the memory and the cache
are coherent before transfering the data.

This is however a costly operation, that in controlled circumstances is not
needed.

Hence the addition of the Clean_Cache parameter to the Copy_Rect* subprograms.

This also adds a missing variant for Set_Pixel_Blend with color as native
representation, and improves the implementation of
Soft_Drawing_Bitmap.Copy_Rect to ease the implementation of
hardware-accelerated drawing primitives. STM32.DMA2D_Bitmap is simplified
accordingly.