Touch_Panel: provide more information on the touch events to the user
This allows in particular in multi-touch panels to keep track of individual
fingers.

Also this improves the performance of reading the touch events on the FT*
drivers, by using batch reading of the touch registers.

Finally this commit also fixes the ft6x06 driver implemented on the F469 and
F769 disco boards.