MicroBit: Fix button B
As discussed in #267, MicroBit.Time was configuring the Low Frequency
Clock to use an external crystal, but there is no LFC external crystal
on the micro:bit. Also the button B is connected to the pin used for the
LFC crystal, so this configuration meant the pin was actually not
available as GPIO for the button code.

MicroBit.Time now uses synthesized clock for LFC so the pin is available
for button B.