Optional style checks and warning as error switches
This patch makes style checks and warning as error switches controllable
with a scenario variable: ADL_BUILD_CHECKS (disabled by default).

With this we can enforce style checks on all contributions by enabling
ADL_BUILD_CHECKS in testsuite runs and CI builds.

On the other hand the checks are disabled by default which means it is
easier and more friendly to play with the library and the examples.

Part of #271